backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
scikit-learn
google-api-python-client
textblob
pydantic
requests
shap
google-generativeai
matplotlib